clazz/src/main/java/jvm/clazz/constant/InvokeDynamicConstant.java
package jvm.clazz.constant;

import jvm.clazz.Clazz;
import jvm.clazz.Memory;

public class InvokeDynamicConstant extends Constant {
    private final int bootstrapMethodAttributeIndex;
    private final int nameAndTypeIndex;

    public InvokeDynamicConstant(final Clazz clazz, final Memory memory) {
        super(clazz,Tag.InvokeDynamic);
        bootstrapMethodAttributeIndex = memory.readShortAsInt();
        nameAndTypeIndex = memory.readShortAsInt();
    }

    public int getBootstrapMethodAttributeIndex() {
        return bootstrapMethodAttributeIndex;
    }

    public int getNameAndTypeIndex() {
        return nameAndTypeIndex;
    }
}
